+++ src/commands/faithful/texture.ts
@@ -4,7 +4,8 @@ import { Client, CommandInteraction, Message, MessageEmbed } from "@client";
 import { getTextureMessageOptions } from "@functions/getTexture";
 import { MessageActionRow, MessageSelectMenu, MessageSelectOptionData } from "discord.js";
 import axios from "axios";
-import { deleteInteraction, imageButtons } from "@helpers/buttons";
+import { imageButtons } from "@helpers/buttons";
+import { MinecraftSorter } from "@helpers/sorter";
 
 export const command: SlashCommand = {
 	servers: ["faithful", "faithful_extra", "classic_faithful"],
@@ -74,7 +75,7 @@ export const command: SlashCommand = {
 			// parsing everything correctly
 			for (let i = 0; i < results.length; i++) {
 				results[i] = {
-					label: `[#${results[i].id}] (${results[i].paths[0].versions[0]}) ${results[i].name}`,
+					label: `[#${results[i].id}] (${results[i].paths[0].versions.sort(MinecraftSorter).reverse()[0]}) ${results[i].name}`,
 					description: results[i].paths[0].name,
 					value: `${results[i].id}__${interaction.options.getString("pack", true)}`,
 				};
@@ -142,13 +143,10 @@ export const command: SlashCommand = {
 		}
 
 		// no results
-		interaction
-			.editReply({
-				content: await interaction.text({
-					string: "Command.Texture.NotFound",
-					placeholders: { TEXTURENAME: `\`${name}\`` },
-				}),
-			})
-			.then((message: Message) => message.deleteButton());
+		interaction.deleteReply()
+		interaction.followUp({
+			content: await interaction.text({ string: "Command.Texture.NotFound", placeholders: { TEXTURENAME: `\`${name}\`` } }),
+			ephemeral: true,
+		})
 	},
 };