+++ src/helpers/interfaces/firestorm/paths.ts
@@ -1,0 +1,8 @@
+export interface Path {
+  id: string;
+  use: string; // use ID
+  name: string; // texture path
+  mcmeta: boolean; // true if animated
+  versions: string[]; // texture versions
+}
+export interface Paths extends Array<Path> {}